Added DirectMessages route & refactored SendMessage component
The SendMessage component has been decoupled from react-apollo and is now to be used only for sending messages. Now it can be user by the ViewTeam route to Send Messages within a channel's chat, and also by the new DirectMessages route to send messages between two users on a team. All Apollo queries should now be done in the parent component of Send MEssage