Configured apollo-upload-client for file uploads
Apollo-upload-client takes the place of the standard createHttpLink and allows us to upload files. Had to change the afterware to get it working, , the headers are now a direct property of operation.getContext(), and are stored in an object. So just changed how I was retrieving the tokens from the headers