Implemented 'offset pagination' to load more messages on button click
This method isn't the best for the use case but is easiest to implement. A new parameter to the messages query accepts offset, which is where the query begins on the list of messages. Starting at offset 0, we fetch a limited # of messages from the back end (currently 35). Then, a Load More button is clicked to use Apollo's fetchMore method and to perform another query with an offset starting at the # of already loaded messages. We then update the local messages cache, and stop rendering the Load More button when no new messages are returned from the server