Implemented 'Infinite Scrolling' for loading old messages
Completely replaced the Load More button with implicit loading when the user scrolls to the top (oldest messages) of a message feed. To do this, we ref the div holding the messages, then keep track of the scrollTop to know when the user has neared the top. At this point, perform a new cursor paginated query to get the next batch of messages from the server, then recalculate height scroll position so they dont get bumped to a weird spot. Using internal state reachedTop: boolean to ensure pagination query doesnt get called more than once