Fixed subscription issue on logins
Before, we had to refresh the app after logins to get subscriptions correctly. Now, we enabled the lazy option for WebSocketLink, and turned connectionParams into a function that will run and return the params (rather than setting them directly). This enables us to explicitly call redirect in the Login route after a successful login by importing the wsLink instance we created and calling wsLink.subscriptionClient.tryReconnect()